hw_14/requirements.txt
google
protobuf>=6.33.2
python-memcached
